requirements_dev.txt
invoke
twine
wheel
zest.releaser
django-model-utils>=2.0
